+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 * ADDED: Support for paging with page and per_page
 * ADDED: User defined location for the static files (including SPAs)
 * ADDED: Console parameter for help and version
-
+* FIXED: Adding to the empty collection will use the value of the id-field if it is set
+ 
 ### [0.8.0] - 2018-10-14
 * Release as a dotnet tool
 * Use executable location as a base path for settings files